ingestion/finnhub-websocket-prod_send/requirements.txt
azure-eventhub
finnhub-python==2.4.27
websocket-client==1.9.0
websockets==15.0.1

fastapi==0.115.0
uvicorn[standard]==0.30.0
